added jwt auth api

+++ src/services/posts.py
@@ -7,7 +7,7 @@ from db.database import get_db
 from db.models.post import Post
 from schemas.base import ListQuery
 from schemas.post import PostSchema
-from services.exceptions import NotFound
+from services.exceptions import NotFound, NotAllowed
 
 
 class PostsService:
@@ -23,30 +23,40 @@ class PostsService:
         return item
 
     async def read(self, query: ListQuery) -> list[Post]:
-        # TODO: offset
         posts = await self.db.execute(
             select(
                 Post.id,
-                Post.text
+                Post.text,
+                Post.user_id
+            ).where(
+                Post.id >= query.offset
             ).limit(query.limit)
         )
         return posts.all()
 
-    async def create(self, post: PostSchema) -> Post:
+    async def create(self, user_id: int, post: PostSchema) -> Post:
         db_post = Post(**post.dict())
+        db_post.user_id = user_id
         self.db.add(db_post)
         await self.db.commit()
         return db_post
 
-    async def update(self, post_id: int, post: PostSchema) -> Post:
+    async def update(self, user_id: int, post_id: int, post: PostSchema) -> Post:
         db_post = await self.get(post_id)
+        if db_post.user_id != user_id:
+            raise NotAllowed
+
+        post.user_id = user_id
         db_post.update(**post.dict(exclude_none=True))
 
         await self.db.commit()
         return db_post
 
-    async def delete(self, post_id: int) -> None:
+    async def delete(self, user_id: int, post_id: int) -> None:
         db_post = await self.get(post_id)
+        if db_post.user_id != user_id:
+            raise NotAllowed
+
         await self.db.delete(db_post)
         await self.db.commit()
